% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire damage data set}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{distance}{fire source distance}
  \item{damage}{damage from fire source}
  ...
}
}
\source{
na
}
\usage{
fire
}
\description{
A set of data for damage caused by fire
Report ...
}
\keyword{datasets}
