% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{The myci Function}
\usage{
myci(x)
}
\arguments{
\item{x}{A numeric vector representing the sample data. The function will calculate the sample mean and standard deviation.}
}
\value{
A numeric vector of length 2 containing the lower and upper bounds of the 95% confidence interval for the population mean.
The vector is named "Lower" and "Upper" for the respective bounds.
}
\description{
This function calculates the 95% confidence interval for the population mean mu based on a single sample.
It uses the t-distribution and assumes the sample size is less than 30 or that the population variance is unknown.
}
\examples{
# Example usage:
set.seed(23)
x <- rnorm(30, mean = 10, sd = 12)  # Generate a sample
myci(x)  # Call the myci function to get the 95\% CI

}
