% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My CLT Poisson Distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Integer. The sample size for each iteration (i.e., the number of random values drawn from the Poisson distribution in each iteration).}

\item{iter}{Integer. The number of iterations to perform. In each iteration, a new random sample of size `n` is drawn, and the sample mean is calculated.}

\item{lambda}{Numeric. The rate parameter (lambda) for the Poisson distribution. The default value is 10.}

\item{...}{Additional graphical parameters passed to the `hist` function for customizing the histogram of the sample means.}
}
\value{
This function does not return a value but produces three plots:
\itemize{
  \item A histogram of the sample means with an overlaid theoretical normal curve.
  \item A barplot showing the relative frequencies of the sampled Poisson values.
  \item A plot of the theoretical Poisson probability mass function.
}
}
\description{
This function demonstrates the Central Limit Theorem (CLT) using a Poisson distribution. It simulates random samples from a Poisson distribution with a specified lambda parameter, calculates the sample means over multiple iterations, and creates graphical representations of the results. These include a histogram of the sample means, a barplot of the relative frequencies of the Poisson-distributed samples, and a plot of the theoretical Poisson probability mass function.
}
\examples{
# Example usage of the function with n = 10, iter = 1000, lambda = 5
mycltp(n = 10, iter = 1000, lambda = 5)
}
