% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A function representing the likelihood function to be evaluated. It should take
parameters in the form of `lfun(x, param)` where `x` is a vector of data points
and `param` is a parameter value.}

\item{x}{A numeric vector of data points for which the likelihood function will be evaluated.}

\item{param}{A numeric vector of parameter values at which the likelihood function will be evaluated.}

\item{...}{Additional arguments to be passed to the plotting function (e.g., main, xlab, ylab).}
}
\value{
A list containing:
        - `i`: The index of the maximum likelihood estimate.
        - `parami`: The parameter value that maximizes the likelihood.
        - `yi`: The maximum likelihood value at `parami`.
        - `slope`: The estimated slope of the likelihood function at the maximum, or "NA" if not computable.
}
\description{
This function performs maximum likelihood estimation for a given likelihood function
and plots the resulting likelihood values against the specified parameters.
}
\examples{
# Example usage of the mymaxlik function with a sample likelihood function
sample_likelihood = function(x, theta) {
  dnorm(x, mean = theta, sd = 1)
}
data_points = rnorm(100, mean = 5, sd = 2)
results = mymaxlik(sample_likelihood, data_points, param = seq(0, 10, length = 100))

}
