% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{The Sample Function}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Integer. The number of samples to draw in each iteration.}

\item{iter}{Integer. The number of iterations to perform.}

\item{time}{Numeric. The amount of time (in seconds) to pause between iterations.}
}
\value{
This function does not return a value. It produces a series of bar plots displaying the distribution of the sampled data.
}
\description{
The Sample Function
}
\examples{
# Example usage of the mysample function:
mysample(n = 20, iter = 5, time = 1)
}
